fix: implement the user permission  to restrict the shop related server actions and pages.

+++ actions/shopActions.ts
@@ -1,6 +1,6 @@
 'use server'
 
-import dbConnect from "@/dbConnect";
+import { getServerSessionContext } from "@/lib/checkServerAuth";
 import Shop from "@/models/shopModel";
 import { revalidatePath } from "next/cache";
 
@@ -14,9 +14,13 @@ export async function updateShop(domain: string, data: UpdateData) {
     if (!domain || !data) {
         return { success: false, message: "Domain or updated data is missing." };
     }
+
     try {
-        //connect to the database
-        await dbConnect();
+        const { role } = await getServerSessionContext();
+
+        if (role !== 'merchant') {
+            return { success: false, message: "Unauthorized. Only admin can update shops." };
+        }
 
         //find the shop by domain and update the name
         const result = await Shop.findOneAndUpdate({ domain }, { $set: data });
@@ -25,26 +29,25 @@ export async function updateShop(domain: string, data: UpdateData) {
             return { success: false, message: "Update Shop failed. No shop found with the provided domain." };
         }
 
+        return { success: true, message: "Shop updated successfully." };
         // eslint-disable-next-line @typescript-eslint/no-explicit-any
     } catch (error: any) {
         console.log("Error updating shop:", error.message);
         return { success: false, message: error.message };
     }
-
-    return { success: true, message: "Shop updated successfully." };
 }
 
-export async function getShops(user: string) {
-    if (!user) {
-        console.log("User Id is required to get the connected shops.");
-        return [];
-    }
-
+export async function getShops() {
     try {
-        await dbConnect();
+        const { role, merchantId } = await getServerSessionContext();
+
+        if (role === 'driver') {
+            console.log("Driver role detected. No shops to fetch.");
+            return [];
+        }
 
         // Fetch shops associated with the user
-        const shops = await Shop.find({ user });
+        const shops = await Shop.find({ owner: merchantId });
 
         if (shops.length === 0) {
             console.log("No shops found for the user.");
@@ -52,10 +55,10 @@ export async function getShops(user: string) {
         }
 
         //format shops for removing unnecessary fields
-        const formattedShops = shops.map(shop=>({
-                domain: shop.domain,
-                platform: shop.platform,
-                name: shop.name  
+        const formattedShops = shops.map(shop => ({
+            domain: shop.domain,
+            platform: shop.platform,
+            name: shop.name
         }))
 
         return formattedShops;
@@ -73,7 +76,11 @@ export async function deleteShop(domain: string) {
     }
 
     try {
-        await dbConnect();
+        const { role } = await getServerSessionContext();
+
+        if (role !== 'merchant') {
+            return { success: false, message: "Unauthorized. Only admin can delete shops." };
+        }
 
         const result = await Shop.findOneAndDelete({ domain });
